+++ src/main/java/ru/thevalidator/daivinchikmatcher2/config/settings/options/BaseSettings.java
@@ -1,4 +1,4 @@
-package ru.thevalidator.daivinchikmatcher2.config.settings;
+package ru.thevalidator.daivinchikmatcher2.config.settings.options;
 
 public interface BaseSettings {
     public void setLongPollWaitTime(Integer delayInSeconds);
