+++ src/main/java/ru/thevalidator/daivinchikmatcher2/task/Task.java
@@ -1,4 +1,4 @@
-package ru.thevalidator.daivinchikmatcher2.vk.task;
+package ru.thevalidator.daivinchikmatcher2.task;
 
 public interface Task extends Runnable {
     public void stop();
