+++ src/main/java/ru/thevalidator/daivinchikmatcher2/task/longpoll/UserLongPollSpectatorService.java
@@ -1,4 +1,4 @@
-package ru.thevalidator.daivinchikmatcher2.vk.task.longpoll;
+package ru.thevalidator.daivinchikmatcher2.task.longpoll;
 
 import com.vk.api.sdk.client.ClientResponse;
 import com.vk.api.sdk.client.VkApiClient;
@@ -13,7 +13,7 @@ import ru.thevalidator.daivinchikmatcher2.config.settings.Settings;
 import ru.thevalidator.daivinchikmatcher2.util.data.SerializerUtil;
 import ru.thevalidator.daivinchikmatcher2.vk.custom.actor.CustomUserActor;
 import ru.thevalidator.daivinchikmatcher2.vk.dto.LongPollServerResponse;
-import ru.thevalidator.daivinchikmatcher2.vk.task.Task;
+import ru.thevalidator.daivinchikmatcher2.task.Task;
 
 import java.io.IOException;
